Midterm/requirements.txt
chainlit==1.0.506
cohere==4.37
openai==1.25.1
python-dotenv==1.0.0
langchain==0.1.17
langchain-core==0.1.49
langchain-community==0.0.36
langchain-openai==0.1.5
qdrant-client==1.9.0
tiktoken==0.6.0
pymupdf==1.24.2